% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexplorer.R
\name{dataexplorer}
\alias{dataexplorer}
\title{Play with ilostat data explorer}
\usage{
dataexplorer()
}
\description{
open ilostat data explorer app on your computer.
}
\examples{
\dontrun{

require(shiny)

 dataexplorer()

}
}
\references{
See citation("Rilostat")
}
\author{
David Bescond \email{bescond@ilo.org}
}
