% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquery.R
\name{dataquery}
\alias{dataquery}
\title{Play with ilostat micro data query}
\usage{
dataquery()
}
\description{
open ilostat micro data query app on your computer.
}
\examples{
\dontrun{

require(shiny)

 dataquery()

}
}
\references{
See citation("Rilostat")
}
\author{
David Bescond \email{bescond@ilo.org}
}
