% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilostat.R, R/utils.R
\name{ilostat}
\alias{ilostat}
\title{Get more Support}
\usage{
ilostat(id = NULL)
}
\arguments{
\item{id}{a character, file that to use if \code{NULL} return the list of all available file.}
}
\description{
On regular basis new tutorial and examples are building and available through this function.
}
\details{
brief description of the package

\tabular{ll}{
Package: \tab Rilostat\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2015-2017\cr
License: \tab BSD_2_clause + LICENSE\cr
LazyLoad: \tab yes\cr
}

R Tools for ilostat Open Data
}
\examples{
\dontrun{

# check which documentation have been recently added:

ilostat() 

# get the first vignette:

ilostat('RilostatStart') 

}
}
\references{
See citation("Rilostat")

ilostat bulk download facility user guidelines 
\url{https://www.ilo.org/ilostat-files/Documents/ILOSTAT_BulkDownload_Guidelines.pdf}
}
\author{
David Bescond \email{bescond@ilo.org}
}
\keyword{package}
