% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rilostat.R
\docType{package}
\name{Rilostat}
\alias{Rilostat-package}
\alias{Rilostat}
\title{R Tools for ilostat Open Data}
\description{
This package provides a user-friendly interface to access, download, and manage ILOSTAT open data.
}
\details{
On a regular basis new tutorials and examples are built and available through this function.

\tabular{ll}{
Package: \tab Rilostat\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2020-2025\cr
License: \tab BSD_2_clause + LICENSE\cr
LazyLoad: \tab yes\cr
}
}
\examples{
\dontrun{
# check which documentation have been recently added:
# help(Rilostat)

# https://ilostat.github.io/Rilostat/
}
}
\references{
See citation("Rilostat")

ilostat bulk download facility user guidelines
\url{https://webapps.ilo.org/ilostat-files/Documents/ILOSTAT_BulkDownload_Guidelines.pdf}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ilostat.github.io/Rilostat/}
  \item Report bugs at \url{https://github.com/ilostat/Rilostat/issues}
}

}
\author{
David Bescond \email{bescond@ilo.org}
}
\keyword{internal}
