% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExplorer.R
\name{runExplorer}
\alias{runExplorer}
\title{play with ilostat explorer}
\usage{
runExplorer(display.mode = "normal")
}
\description{
open ilostat explorer app on your computer.
}
\examples{
\dontrun{

 runExplorer()

}
}
\references{
See citation("Rilostat")
}
\author{
David Bescond \email{bescond@ilo.org}
}
