% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rilostat.R
\name{Rilostat}
\alias{Rilostat}
\alias{ilostat_url}
\title{Get more Support}
\usage{
ilostat_url()
}
\description{
On regular basis new tutorial and examples are building and available through this function.
}
\details{
brief description of the package

\tabular{ll}{
Package: \tab Rilostat\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2020-2025\cr
License: \tab BSD_2_clause + LICENSE\cr
LazyLoad: \tab yes\cr
}

R Tools for ilostat Open Data
}
\examples{
\dontrun{

# check which documentation have been recently added:

 # help(Rilostat) 

 # https://ilostat.github.io/Rilostat/ 

}
}
\references{
See citation("Rilostat")

ilostat bulk download facility user guidelines 
\url{https://webapps.ilo.org/ilostat-files/Documents/ILOSTAT_BulkDownload_Guidelines.pdf}
}
\author{
David Bescond \email{bescond@ilo.org}
}
\keyword{package}
