% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExplorer.R
\name{runExplorer}
\alias{runExplorer}
\title{Play with ilostat explorer}
\usage{
runExplorer(display.mode = "normal")
}
\arguments{
\item{display.mode}{The mode in which to display the application. 
If set to the value "showcase", shows application code and metadata 
from a DESCRIPTION file in the application directory alongside the application. 
If set to "normal", displays the application normally. Defaults to "auto", which 
displays the application in the mode given in its DESCRIPTION file, if any.}
}
\description{
open ilostat explorer app on your computer.
}
\examples{
\dontrun{

 runExplorer()

}
}
\references{
See citation("Rilostat")
}
\author{
David Bescond \email{bescond@ilo.org}
}
